+++ src/main/java/com/beauty4u/backend/file/command/application/dto/FileDeleteReqDTO.java
@@ -1,19 +1,16 @@
-package com.beauty4u.backend.basesystem.command.application.dto;
+package com.beauty4u.backend.file.command.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class FileSaveReqDTO {
+public class FileDeleteReqDTO {
 
     private Long entityId;
-    private List<String> imageUrls;
     private Boolean isInform;
 }
