+++ src/main/java/com/beauty4u/backend/file/command/domain/aggregate/FileInfo.java
@@ -1,4 +1,4 @@
-package com.beauty4u.backend.basesystem.command.domain.aggregate;
+package com.beauty4u.backend.file.command.domain.aggregate;
 
 import com.beauty4u.backend.inform.command.domain.aggregate.Inform;
 import com.beauty4u.backend.inquiry.command.domain.aggregate.Inquiry;
