+++ src/main/java/com/beauty4u/backend/file/command/domain/repository/FileRepository.java
@@ -1,6 +1,6 @@
-package com.beauty4u.backend.basesystem.command.domain.repository;
+package com.beauty4u.backend.file.command.domain.repository;
 
-import com.beauty4u.backend.basesystem.command.domain.aggregate.FileInfo;
+import com.beauty4u.backend.file.command.domain.aggregate.FileInfo;
 import com.beauty4u.backend.inform.command.domain.aggregate.Inform;
 import com.beauty4u.backend.inquiry.command.domain.aggregate.Inquiry;
 import org.springframework.data.jpa.repository.JpaRepository;
